use hexagonal architecture
