+++ project/apps/question-service/src/questions.module.ts
@@ -1,7 +1,10 @@
 import { Module } from '@nestjs/common';
-import { QuestionsController } from './questions.controller';
-import { QuestionsService } from './questions.service';
 import { ConfigModule } from '@nestjs/config';
+
+import { QuestionsController } from 'src/adapters/controllers/questions.controller';
+import { QuestionsService } from 'src/domain/ports/questions.service';
+import { QuestionsRepository } from 'src/domain/ports/questions.repository';
+import { SupabaseQuestionsRepository } from 'src/adapters/db/questions.supabase';
 
 @Module({
   imports: [
@@ -10,6 +13,12 @@ import { ConfigModule } from '@nestjs/config';
     }),
   ],
   controllers: [QuestionsController],
-  providers: [QuestionsService],
+  providers: [
+    QuestionsService,
+    {
+      provide: QuestionsRepository,
+      useClass: SupabaseQuestionsRepository,
+    },
+  ],
 })
 export class QuestionsModule {}
